+++ CHANGELOG.md
@@ -11,6 +11,9 @@ otherwise.
 
 * `plookupAssoc`, designed as a very general treatment of list-like structures
   as assoc lists.
+* `Plutarch.Extra.Ord` module, containing a new type `POrdering` corresponding
+  to the Haskell `Ordering`, as well as `PComparator`, representing a
+  \'materialized ordering\'.
 
 ### Removed